requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
openai==0.10.2
